rpi4_64_router: bump linux version; fixes; add wireguard tools
Bump the Linux version from 5.10.110 to 5.15.80.

As a result of the version bump, the default kernel config being used
now compresses the kernel modules using XZ. This necessitated changing
the host kmod package to have support for XZ compression, as host kmod
is used to generate modules.dep for the target. Also, the busybox
version of kmod doesn't support compressed modules, so add the kmod
package and kmod tools to have a modprobe command that can handle XZ
compressed modules.

Adds the wireguard tools package for potential future use.

Note that we previously jumped to Linux 5.15, then back to 5.10 because
it looked like there were firmware problems with one of the USB3-ETH
adapters I use in my setup (see commit
81a44d1305de191ef64b6f23c1595cbbcbdc6a2e). As of now my Internet seems
to be working fine on all USB3-ETH interfaces, though I will note that
the dmesg output does display the following that I'm almost positive
wasn't in 5.10:

[    3.143582] r8152 2-2:1.0: Direct firmware load for rtl_nic/rtl8153a-4.fw failed with error -2
[    3.157642] r8152 2-2:1.0: unable to load firmware patch rtl_nic/rtl8153a-4.fw (-2)

I don't know if this is a problem to be concerned about. I will note
that usb 2-2 is:

[    2.918292] usb 2-2: new SuperSpeed USB device number 3 using xhci_hcd
[    2.951902] usb 2-2: New USB device found, idVendor=0bda, idProduct=8153, bcdDevice=30.00
[    2.960178] usb 2-2: New USB device strings: Mfr=1, Product=2, SerialNumber=6
[    2.967356] usb 2-2: Product: USB 10/100/1000 LAN
[    2.972145] usb 2-2: Manufacturer: CMI
[    2.975936] usb 2-2: SerialNumber: 000001
[    3.110785] usb 2-2: reset SuperSpeed USB device number 3 using xhci_hcd

Signed-off-by: Joel Carlson <JoelsonCarl@gmail.com>